Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0029-divide-two-integers/0029-divide-two-integers.cpp
@@ -8,8 +8,8 @@ public:
         else if (divisor <= 0 && dividend > 0)
             sign = false;
 
-        int n = abs(dividend);
-        int d = abs(divisor);
+        long n = abs((long)dividend);
+        long d = abs((long)divisor);
 
         long ans = 0;
